added conditional device selection with settings and create new features
