+++ src/Components/DeviceItem.js
@@ -1,17 +1,34 @@
 import {useState , useEffect} from "react";
 
-export default function DeviceItem({devices, index, setDevice, device}) {
+export default function DeviceItem({devices, index, setDevice, device, setAddDeviceToggle, setSettingsToggle, from, settingsToggle}) {
 
     const [colorStyle, setColorStyle] = useState("device-line");
 
     const handleDeviceClick = () => {
+
+        if (from === "Settings"){
+            setSettingsToggle(false);
+            setAddDeviceToggle(false);
+        }
+
         setDevice(devices);
     }
     
     useEffect(() =>{
+
         if (Object.keys(device).length !== 0){
             if(device.cat_num === devices.cat_num){
+
+                if (settingsToggle){
+                    setColorStyle("device-line");
+                    return
+                }
+
                 setColorStyle("device-line selected");
+                if (from === "Performance"){
+                    setSettingsToggle(false);
+                    setAddDeviceToggle(false);
+                }
             } else {
                 setColorStyle("device-line");
             }
@@ -21,7 +38,7 @@ export default function DeviceItem({devices, index, setDevice, device}) {
             }
         }
 
-    },[device])
+    },[device,settingsToggle])
 
     return (
         <li>
