+++ src/Containers/PerformanceView.js
@@ -19,9 +19,6 @@ import { styled } from '@mui/material/styles';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Switch from '@mui/material/Switch';
 import Button from "../Components/Button";
-import { useAuth} from "../Provider/authProvider";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
 
 const IOSSwitch = styled((props) => (
     <Switch focusVisibleClassName=".Mui-focusVisible" disableRipple {...props} />
@@ -125,35 +122,13 @@ const dataset = [
     },
   ];
 
-export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, setUser}) {
+export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, handlePlantPalClick, handleRefreshClick, devices, lastLog, device, setDevice, refreshDate, handleLogout, settingsToggle}) {
 
     const [autoSwitch, setAutoSwitch] = useState(false);
-    const [devices, setDevices] = useState([]);
-    const [device, setDevice] = useState({});
-    const [deviceLogs, setDeviceLogs] = useState([]);
-    const [lastLog, setLastLog] = useState({});
     const [moistureLevel, setMoistureLevel] = useState(0);
     const [waterStatus, setWaterStatus] = useState("");
     const [waterStatusCSS, setWaterStatusCSS] = useState("good-water");
     const [waterStatusImg, setWaterStatusImg] = useState("");
-    const [refreshDate, setRefreshDate] = useState("");
-
-    const navigate = useNavigate();
-    const { clearToken, token } = useAuth();
-
-    const client = axios.create({
-        baseURL: process.env.REACT_APP_BASE_URL,
-        
-        headers: {
-            "Authorization": "Bearer " + token
-        }
-        
-    });
-
-    const handleLogout = () => {
-        clearToken();
-        navigate("/auth", { replace: true });
-    };
 
     const handleAutoSwitch = (e) => {
         setAutoSwitch(e.target.checked);
@@ -166,41 +141,6 @@ export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, 
     const handleAddDeviceClick = () => {
         setAddDeviceToggle(true);
         setSettingsToggle(true);
-    }
-
-    const handlePlantPalClick = () => {
-        navigate('/', {
-            replace: true,
-        });
-    }
-
-    const fetchUserDevices = async () => {
-        try {
-            const response = await client.get("/dashboard/userDevices");
-
-            setDevices(response.data.devices);
-            
-        } catch (error) {
-            clearToken();
-            navigate("/auth", { replace: true });
-        }
-    }
-
-    const handleRefreshClick = () => {
-        fetchUserDevices();
-    }
-
-    const fethDeviceLogs = async () => {
-        try {
-            const response = await client.get("/dashboard/deviceLogs", { params: {cat_num: device.cat_num}});
-
-            setDeviceLogs(response.data.deviceLogs);
-            setLastLog(response.data.lastLog);
-            
-        } catch (error) {
-            clearToken();
-            navigate("/auth", { replace: true });
-        }
     }
 
     const formatMostureLevel = () => {
@@ -239,27 +179,6 @@ export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, 
 
     }
 
-    const formatRefreshDate = () => {
-        const userTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-        const dateObject = new Date(lastLog.log_date);
-
-        const localTimeString = dateObject.toLocaleString(undefined, {
-            timeZone: userTimeZone,
-        });
-          
-        setRefreshDate(localTimeString);
-    }
-
-    useEffect(() => {
-        fetchUserDevices();
-    }, [device]);
-
-    useEffect(() => {
-        if(typeof device !== "undefined"){
-            fethDeviceLogs();
-        }
-    }, [device]);
-
     useEffect(() => {
         if(typeof lastLog !== "undefined"){
             formatMostureLevel();
@@ -268,14 +187,6 @@ export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, 
 
     useEffect(() => {
         formatWaterStatus();
-    }, [lastLog]);
-
-    useEffect(() => {
-        if(typeof lastLog !== "undefined"){
-            formatRefreshDate();
-        } else {
-            setRefreshDate("");
-        }
     }, [lastLog]);
 
     return (
@@ -308,7 +219,7 @@ export default function PerformanceView({setSettingsToggle, setAddDeviceToggle, 
                     
                     <ul className='device-list'>
                         
-                        { devices.map((devices, index) => <DeviceItem key={devices.device_id} devices={devices} index={index} setDevice={setDevice} device={device}></DeviceItem>)}
+                        { devices.map((devices, index) => <DeviceItem key={devices.device_id} devices={devices} index={index} setDevice={setDevice} device={device} setAddDeviceToggle={setAddDeviceToggle} setSettingsToggle={setSettingsToggle} from="Performance" settingsToggle={settingsToggle}></DeviceItem>)}
 
                     </ul>
